+++ temba/classifiers/types/luis/type.py
@@ -34,19 +34,18 @@ class LuisType(ClassifierType):
     You can find the attributes for your app on your Luis.ai app page.
     """
 
-    @classmethod
-    def get_active_intents_from_api(cls, classifier, logs):
+    def get_active_intents_from_api(self, classifier, logs):
         """
         Gets the current intents defined by this app, in LUIS that's an attribute of the app version
         """
-        app_id = classifier.config[cls.CONFIG_APP_ID]
-        version = classifier.config[cls.CONFIG_VERSION]
-        endpoint_url = classifier.config[cls.CONFIG_ENDPOINT_URL]
-        primary_key = classifier.config[cls.CONFIG_PRIMARY_KEY]
+        app_id = classifier.config[self.CONFIG_APP_ID]
+        version = classifier.config[self.CONFIG_VERSION]
+        endpoint_url = classifier.config[self.CONFIG_ENDPOINT_URL]
+        primary_key = classifier.config[self.CONFIG_PRIMARY_KEY]
 
         start = timezone.now()
         url = endpoint_url + "/apps/" + app_id + "/versions/" + version + "/intents"
-        response = requests.get(url, headers={cls.AUTH_HEADER: primary_key})
+        response = requests.get(url, headers={self.AUTH_HEADER: primary_key})
         elapsed = (timezone.now() - start).total_seconds() * 1000
 
         log = HTTPLog.from_response(HTTPLog.INTENTS_SYNCED, url, response, classifier=classifier)