+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v5.3.4
+----------
+ * Release flow starts when releasing flows
+
+v5.3.3
+----------
+ * Add releasing to Classifiers and HTTPLogs
+
+v5.3.2
+----------
+ * Allow manual syncing of classifiers
+
 v5.3.1
 ----------
  * Update documentation for FB webhook events to subscribe to